memcpy--tolower

+++ libft/ft_memset.c
@@ -6,10 +6,12 @@
 /*   By: cboudrin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/23 10:34:58 by cboudrin          #+#    #+#             */
-/*   Updated: 2021/11/23 14:55:58 by cboudrin         ###   ########.fr       */
+/*   Updated: 2021/11/24 11:12:35 by cboudrin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-void	*ft_memset(void *p, int v, unsigned int len)
+#include "libft.h"
+
+void	*ft_memset(void *p, int v, size_t len)
 {
 	unsigned char	*i;
